pyproject.toml
[project]
name = "orionbelt-runner"
version = "0.1.0"
description = "Run OrionBelt Semantic Layer query batches on a schedule and emit reports."
license = {text = "BSL-1.1"}
authors = [{name = "Ralf Becher, RALFORION d.o.o.", email = "ralf.becher@web.de"}]
requires-python = ">=3.12"
readme = "README.md"
keywords = [
    "semantic-layer",
    "orionbelt",
    "reporting",
    "scheduled-queries",
]
dependencies = [
    "pydantic>=2.7",
    "httpx>=0.27",
    "ruamel-yaml>=0.18",
    "typer>=0.12",
    "structlog>=24.1",
]

[project.scripts]
orionbelt-runner = "orionbelt_runner.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "respx>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orionbelt_runner"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
